[#666] Extend Xtext Generator for JUnit 5
- Added enum JUnitVersion
- Create JUnitFragment from Junit4Fragment2
  - provide setter for JUnit version
  - Deprecated Junit4Fragment2 and extend from JUnitFragment
  - changed StandardLanguage#junitSupport type to JUnitFragment
- Extended JUnitFragment to produce JUnit version specific code
- Added junitVersion property to WizardConfiguration
- generate JUnit 5 specific configuration to workflow when JUnit 5 is
configured
- TestProjectDescriptor: Configure JUnit 5 dependencies
