+++ org.eclipse.xtext.xtext.generator/xtend-gen/org/eclipse/xtext/xtext/generator/StandardLanguage.java
@@ -29,7 +29,7 @@ import org.eclipse.xtext.xtext.generator.grammarAccess.GrammarAccessFragment2;
 import org.eclipse.xtext.xtext.generator.idea.IdeaPluginGenerator;
 import org.eclipse.xtext.xtext.generator.idea.parser.antlr.XtextAntlrIDEAGeneratorFragment;
 import org.eclipse.xtext.xtext.generator.index.ResourceDescriptionStrategyFragment;
-import org.eclipse.xtext.xtext.generator.junit.Junit4Fragment2;
+import org.eclipse.xtext.xtext.generator.junit.JUnitFragment;
 import org.eclipse.xtext.xtext.generator.parser.antlr.XtextAntlrGeneratorFragment2;
 import org.eclipse.xtext.xtext.generator.resourceFactory.ResourceFactoryFragment2;
 import org.eclipse.xtext.xtext.generator.scoping.ImportNamespacesScopingFragment2;
@@ -114,7 +114,7 @@ public class StandardLanguage extends XtextGeneratorLanguage {
   
   private ContentAssistFragment2 contentAssist = new ContentAssistFragment2();
   
-  private Junit4Fragment2 junitSupport = new Junit4Fragment2();
+  private JUnitFragment junitSupport = new JUnitFragment();
   
   private RefactorElementNameFragment2 renameRefactoring = new RefactorElementNameFragment2();
   
@@ -417,11 +417,11 @@ public class StandardLanguage extends XtextGeneratorLanguage {
   }
   
   @Pure
-  protected Junit4Fragment2 getJunitSupport() {
+  protected JUnitFragment getJunitSupport() {
     return this.junitSupport;
   }
   
-  public void setJunitSupport(final Junit4Fragment2 junitSupport) {
+  public void setJunitSupport(final JUnitFragment junitSupport) {
     this.junitSupport = junitSupport;
   }
